+++ src/api/meetings.ts
@@ -8,7 +8,6 @@ import {
   MyMeetingsResponse,
   FavoritesResponse,
 } from "@/types";
-import type { MeetingParticipantsResponse } from "@/types/meetingTypes";
 
 export async function getMeetingList(
   params: GetMeetingListParams,
