+++ src/app/meetings/modal/MeetingCategoryStep.tsx
@@ -9,14 +9,8 @@ import studyImage from "@/assets/img/category/study.jpg";
 import towerWork from "@/assets/img/category/business.jpg";
 import JobIShoes from "@/assets/img/category/fitness-health.jpg";
 import etcImage from "@/assets/img/category/etc.jpg";
-import { MeetingCategoryStepProps } from "@/types/meeting/meeting-form.props";
-
-export interface MeetingCategoryItem {
-  value: string;
-  label: string;
-  imageSrc: StaticImageData;
-  className?: string;
-}
+import { MeetingCategoryStepProps } from "@/types/meeting/meetingTypes";
+import { MeetingCategoryItem } from "@/types/meeting/meetingTypes";
 
 export const MEETING_CATEGORY_LIST: MeetingCategoryItem[] = [
   {
