--- src/types/meeting/meeting-image.types.ts
@@ -1,4 +1,0 @@
-export interface UploadImageResponse {
-  presignedUrl: string;
-  publicUrl: string;
-}
