+++ src/types/meeting/meetingTypes.ts
@@ -1,3 +1,5 @@
+import type { StaticImageData } from "next/image";
+
 export interface MeetingDetailHost {
   id: number;
   name: string;
@@ -19,37 +21,6 @@ export interface RecommendedMeetingItem {
   capacity: number;
   registrationEnd: string;
   dateTime: string;
-}
-
-export interface MeetingDetailData {
-  id: number;
-  teamId: string;
-  name: string;
-  type: string;
-  region: string;
-  address: string;
-  link: string;
-  latitude: number;
-  longitude: number;
-  dateTime: string;
-  registrationEnd: string;
-  capacity: number;
-  participantCount: number;
-  image: string | null;
-  description: string;
-  canceledAt: string | null;
-  confirmedAt: string | null;
-  hostId: number;
-  createdBy: number;
-  createdAt: string;
-  updatedAt: string;
-  host: MeetingDetailHost;
-  isFavorited: boolean;
-  isHost: boolean;
-  isJoined: boolean;
-  isLoggedIn: boolean;
-  threads: MeetingThreadItem[];
-  recommendedMeetings: RecommendedMeetingItem[];
 }
 
 export interface MeetingDetailApiData {
@@ -151,4 +122,240 @@ export interface MeetingListResponse {
   data: MeetingListItemApiData[];
   nextCursor: string | null;
   hasMore: boolean;
+}
+
+export interface MeetingDetailData {
+  id: number;
+  teamId: string;
+  name: string;
+  type: string;
+  region: string;
+  address: string;
+  link: string;
+  latitude: number;
+  longitude: number;
+  dateTime: string;
+  registrationEnd: string;
+  capacity: number;
+  participantCount: number;
+  image: string | null;
+  description: string;
+  canceledAt: string | null;
+  confirmedAt: string | null;
+  hostId: number;
+  createdBy: number;
+  createdAt: string;
+  updatedAt: string;
+  host: MeetingDetailHost;
+  isFavorited: boolean;
+  isHost: boolean;
+  isJoined: boolean;
+  isLoggedIn: boolean;
+  threads: MeetingThreadItem[];
+  recommendedMeetings: RecommendedMeetingItem[];
+}
+
+export interface MeetingDetailViewProps {
+  data: MeetingDetailData;
+  participantAvatars: MeetingParticipantUser[];
+  isFavoritePending: boolean;
+  isJoinPending: boolean;
+  isAuthLoading: boolean;
+  actionLabel: string;
+  isActionDisabled: boolean;
+  shouldShowHostMenu: boolean;
+  shouldShowClosedGuide: boolean;
+  canViewLink: boolean;
+  canWriteThread: boolean;
+  linkGuideText: string;
+  threadGuideText: string;
+  onJoin: () => Promise<void> | void;
+  onCancelJoin: () => Promise<void> | void;
+  onAttend: () => Promise<void> | void;
+  onShare: () => Promise<void> | void;
+  onEdit: (nextValues: Partial<MeetingDetailData>) => void;
+  onDelete: () => void;
+  onToggleFavorite: () => void;
+}
+
+export interface MeetingDetailContentProps {
+  meetingId: number;
+  hasAttendedInitially: boolean;
+}
+export interface MeetingDescriptionSectionProps {
+  data: MeetingDetailData;
+}
+export interface MeetingHeaderSectionProps {
+  data: MeetingDetailData;
+  participantAvatars: MeetingParticipantUser[];
+  isFavoritePending: boolean;
+  isJoinPending: boolean;
+  isAuthLoading: boolean;
+  actionLabel: string;
+  isActionDisabled: boolean;
+  shouldShowHostMenu: boolean;
+  shouldShowClosedGuide: boolean;
+  onJoin: () => Promise<void> | void;
+  onCancelJoin: () => Promise<void> | void;
+  onAttend: () => Promise<void> | void;
+  onShare: () => Promise<void> | void;
+  onEdit: (nextValues: Partial<MeetingDetailData>) => void;
+  onDelete: () => void;
+  onToggleFavorite: () => void;
+}
+
+export interface MeetingLinkSectionProps {
+  link: string;
+  canViewLink: boolean;
+  guideText: string;
+}
+
+export interface MeetingThreadSectionProps {
+  meetingId: number;
+  canWriteThread: boolean;
+  guideText: string;
+}
+
+export interface RecommendedMeetingsSectionProps {
+  data: MeetingDetailData;
+}
+
+export interface MeetingDetailPageProps {
+  params: Promise<{
+    id: string;
+  }>;
+}
+
+export interface EditMeetingModalProps {
+  isOpen: boolean;
+  onOpenChange: (isOpen: boolean) => void;
+  data: MeetingDetailData;
+  onSubmit: (nextValues: Partial<MeetingDetailData>) => Promise<void> | void;
+}
+export interface MeetingTypeOption {
+  value: string;
+  label: string;
+}
+
+export interface MeetingCategoryItem {
+  value: string;
+  label: string;
+  imageSrc: StaticImageData;
+  className?: string;
+}
+export interface ScheduleDatePickerProps {
+  id: string;
+  label: string;
+  value: string;
+  hintText?: string;
+  isRequired?: boolean;
+  isDestructive?: boolean;
+  min?: string;
+  max?: string;
+  onChange: (value: string) => void;
+}
+
+export interface ScheduleTimePickerProps {
+  id: string;
+  label?: string;
+  value: string;
+  hintText?: string;
+  isRequired?: boolean;
+  isDestructive?: boolean;
+  onChange: (value: string) => void;
+}
+
+export interface UploadImageResponse {
+  presignedUrl: string;
+  publicUrl: string;
+}
+
+export interface MeetingFormValues {
+  category: string;
+  name: string;
+  description: string;
+  link: string;
+  imageFile: File | null;
+  previewImageUrl: string;
+  imageUrl: string | null;
+  startDate: string;
+  startTime: string;
+  endDate: string;
+  endTime: string;
+  capacity: string;
+}
+
+export interface MeetingFormErrors {
+  category: string;
+  name: string;
+  description: string;
+  link: string;
+  imageUrl: string;
+  startDate: string;
+  startTime: string;
+  endDate: string;
+  endTime: string;
+  capacity: string;
+}
+
+export interface MeetingBasicInfoValues {
+  category?: string;
+  name: string;
+  description: string;
+  link: string;
+  imageFile: File | null;
+  previewImageUrl: string;
+  imageUrl: string | null;
+}
+
+export interface MeetingBasicInfoErrors {
+  category?: string;
+  name: string;
+  description: string;
+  link: string;
+  imageUrl: string;
+}
+
+export interface MeetingScheduleStepValues {
+  startDate: string;
+  startTime: string;
+  endDate: string;
+  endTime: string;
+  capacity: string;
+}
+
+export interface MeetingBasicInfoStepProps {
+  values: MeetingBasicInfoValues;
+  isImageUploading: boolean;
+  errors: MeetingBasicInfoErrors;
+  onChange: (nextValues: {
+    category?: string;
+    name?: string;
+    description?: string;
+    link?: string;
+  }) => void;
+  onChangeImage: (nextFile: File | null) => void;
+  onRemoveImage: () => void;
+}
+
+export interface MeetingBasicInfoSectionProps extends MeetingBasicInfoStepProps {
+  showCategoryField?: boolean;
+  showImageMeta?: boolean;
+}
+
+export interface MeetingCategoryStepProps {
+  value: string;
+  onChange: (value: string) => void;
+}
+
+export interface MeetingScheduleStepProps {
+  values: MeetingScheduleStepValues;
+  errors: {
+    startDate: string;
+    startTime: string;
+    endDate: string;
+    endTime: string;
+    capacity: string;
+  };
+  onChange: (nextValues: Partial<MeetingScheduleStepValues>) => void;
 }
